Deal with separate compilation vs instrumentation
Merged two test instrumentation services to avoid the problem with
separate compilation. So far we deal with it by having a subproject per
instrumentation. This looked like an overkill for those tests.
